+++ csc121-creative-project/src/SeaShells.java
@@ -39,7 +39,7 @@ public class SeaShells {
 	public PApplet draw(PApplet c) {
 		c.noStroke();
 		c.fill(0, 255, 127);  // Sea green color for SeaShells
-		c.ellipse((float)this.loc.getX(), (float)this.loc.getY(), radius * 2, radius * 2);
+		c.ellipse((int)this.loc.getX(), (int)this.loc.getY(), radius * 2, radius * 2);
 		return c;
 	}
